LayoutTests/http/tests/websocket/tests/hybi/contentextensions/display-none-expected.txt

CONSOLE MESSAGE: line 18: onopen
CONSOLE MESSAGE: line 22: onmessage reply sent
CONSOLE MESSAGE: line 25: onclose
This should not be hidden.
